core: missing cast of memory dump disk size to bigint
Recently added upgrade scripts that convert hibernation/memory dump
volumes to disks, were missing the cast mentioned above which causes
an error (that would happen for VMs with more than 1g of memory) of
type: integer out of range

+++ packaging/dbscripts/upgrade/04_00_0040_convert_hibernation_volumes_to_disks.sql
@@ -44,7 +44,7 @@ INSERT INTO images(
 SELECT
     cast(split_part(memory_volume, ',', 4) as uuid),
     'now',
-    (mem_size_mb + 200 + (64 * num_of_monitors)) * 1024 * 1024,
+    (mem_size_mb::bigint + 200 + (64 * num_of_monitors)) * 1024 * 1024,
     '00000000-0000-0000-0000-000000000000',
     1, -- ok
     0, -- unknown (will be updated by the next step)
